Content is dynamic now!
And things have been split apart, because who needs a giant file
anyways?

As of right now, if you have a Jekyll/Mynt/other static-site generator
file that uses yaml frontmatter for config info and markdown for writing
contents, and you put it in the `content` directory, it *should* be read
as a post, and you *should* be able to see it in both the index and
individual post page.

Cool!