It's been awhile...
I sorta just let this sit alone in a corner, instead of actually playing
with it. But that's besides the point. Truth be told, the main reason I
stopped dealing with it before was that I was getting frustrated with
Go's templates. Not coincidentally at all, that's the reason I came
back. Storyteller's going to have a web app soon, and it's probably
going to use Go. It might be nice to know how the templates work.

Anyways, enough about that; changes!

There are quite a few, but mostly in the templates/look and feel
department. Most of the code was "good enough" for what I needed, and
the main focus today was getting the templates working. So, uh,
templates work. That's kinda cool.